requirements.txt
Django>=4.2,<5.0
django-apscheduler
requests
python-dotenv
textblob
transformers
openai
pillow
gunicorn
whitenoise
psycopg2-binary
dj-database-url
channels
channels-redis
